requirements.txt
click
tqdm
numpy
scipy
opencv-python
scikit-image
scikit-learn>=0.18
pytables
pytorch
torchvision
-e git://github.com/Britefury/batchup.git#egg=batchup
